#1529 implemented requested changes refactored SearchResponseUtil methods moved SearchResponseUtil methods to AbstractSearchResponse removed SearchResponseUtil class
